Xong 1 bãi cứt vừa vừa (check trùng&tồn tại)

+++ app/src/main/java/com/example/logincustomer/data/Model/baocao_thuchi.java
@@ -45,7 +45,7 @@ public class baocao_thuchi {
         return sotienthuchi;
     }
 
-    public void setSotienthuchi(int sotienthuchi) {
+    public void setSotienthuchi(double sotienthuchi) {
         this.sotienthuchi = sotienthuchi;
     }
